requirements.txt
Flask==3.0.2
openai==1.16.2
httpx==0.27.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10